+++ feed-page/new-elements.js
@@ -10,33 +10,72 @@ export function createSidebarElement() {
   const spinboxSidebar = document.createElement('article');
   spinboxSidebar.className = 'sidebarModule spinbox-sidebar';
 
-  const title = document.createElement('h4');
-  title.className = 'sidebarHeader__title__webi__style';
-  title.style.padding = '8px 16px';
-  const hiddenTrackCount = document.createElement('span');
-  hiddenTrackCount.id = 'hiddenTracksCount';
-  const hiddenTracksPlural = document.createElement('span');
-  hiddenTracksPlural.id = 'hiddenTracksPlural';
-  hiddenTracksPlural.textContent = 's';
-  title.append(hiddenTrackCount, ' hidden track', hiddenTracksPlural);
+  const spinboxTitle = document.createElement('h4');
+  spinboxTitle.className = 'sidebarHeader sidebarHeader__title__webi__style';
+  const spinboxIcon = document.createElement('img');
+  spinboxIcon.src = chrome.runtime.getURL('images/icon-16.png');
+  spinboxIcon.style.display = 'inline-block';
+  spinboxIcon.style.marginRight = '6px';
+  spinboxIcon.style.verticalAlign = 'bottom';
+  spinboxTitle.append(spinboxIcon, 'Spinbox');
 
   const content = document.createElement('div');
   content.className = 'sidebarContent';
 
-  const recentlyHidden = document.createElement('div');
-  recentlyHidden.id = 'recentlyHiddenTracksContainer';
-  recentlyHidden.style.padding = '8px 16px';
-  const recentlyHiddenTitle = document.createElement('div');
-  recentlyHiddenTitle.innerText = 'Recently hidden';
-  recentlyHiddenTitle.className =
-    'sc-text-secondary sidebarHeader__title__webi__style sc-mb-1x';
+  const errorMessage = document.createElement('div');
+  errorMessage.id = 'spinboxSidebarErrorMessage';
+  errorMessage.className = 'sc-text-error sc-py-1x sc-px-2x';
+
+  const recentlyHiddenContainer = document.createElement('div');
+  recentlyHiddenContainer.id = 'recentlyHiddenTracksContainer';
+  recentlyHiddenContainer.className = 'sc-py-1x sc-px-2x';
+
+  const recentlyHiddenHeading = document.createElement('div');
+  recentlyHiddenHeading.className = 'sc-text-secondary sc-mb-0.5x';
+  recentlyHiddenHeading.style.display = 'flex';
+
+  const recentlyHiddenTitleText = document.createElement('span');
+  recentlyHiddenTitleText.innerText = 'Recently hidden';
+  recentlyHiddenTitleText.className = 'sidebarHeader__title__webi__style ';
+  recentlyHiddenTitleText.style.flexGrow = '1';
+
+  const recentlyHiddenCountSection = document.createElement('span');
+  recentlyHiddenCountSection.className = 'spinbox-text-dim';
+  recentlyHiddenCountSection.style.fontSize = '12px';
+
+  const hiddenTrackCount = document.createElement('span');
+  hiddenTrackCount.id = 'hiddenTracksCount';
+
+  recentlyHiddenCountSection.append(hiddenTrackCount, ' total');
+  recentlyHiddenHeading.append(
+    recentlyHiddenTitleText,
+    recentlyHiddenCountSection
+  );
+
   const recentlyHiddenList = document.createElement('div');
   recentlyHiddenList.id = 'recentlyHiddenTrackList';
-  recentlyHidden.append(recentlyHiddenTitle, recentlyHiddenList);
 
-  content.append(recentlyHidden);
-  spinboxSidebar.append(title, content);
+  recentlyHiddenContainer.append(recentlyHiddenHeading, recentlyHiddenList);
+  content.append(errorMessage, recentlyHiddenContainer);
+  spinboxSidebar.append(spinboxTitle, content);
   return spinboxSidebar;
+}
+
+/**
+ * Create a pull track button
+ * @param onClick
+ * @returns {HTMLButtonElement}
+ */
+export function createPullTrackButton(onClick) {
+  const button = document.createElement('button');
+  button.className =
+    'spinbox-pull sc-button sc-button-medium sc-button-secondary';
+  button.style.marginLeft = '4px';
+  button.innerText = 'Pull';
+  button.onclick = onClick;
+  button.ariaLabel = 'Pull Track';
+  button.title = 'Pull Track';
+  return button;
 }
 
 /**
@@ -51,7 +90,6 @@ export function createHideTrackButton(onClick) {
   hideButton.style.marginLeft = '4px';
   hideButton.innerText = '✕';
   hideButton.onclick = onClick;
-
   hideButton.ariaLabel = 'Hide Track';
   hideButton.title = 'Hide Track';
   return hideButton;
@@ -99,7 +137,8 @@ export function createRecentlyHiddenTrackElement(track, undoHideTrackFn) {
 
   const hiddenTrackDescriptionArtist = document.createElement('div');
   hiddenTrackDescriptionArtist.className =
-    'spinbox-hidden-track-description-artist';
+    'spinbox-hidden-track-description-artist spinbox-text-dim';
+
   if (track.reposterName) {
     hiddenTrackDescriptionArtist.append(track.reposterName);
     hiddenTrackDescriptionArtist.append(createRepostSvg());
@@ -126,4 +165,11 @@ function createRepostSvg() {
   const template = document.createElement('template');
   template.innerHTML = repostSVG;
   return template.content.firstElementChild;
+}
+
+export function createCouldNotFindPlaylistMessage(playlistName) {
+  const message = document.createElement('div');
+  message.className = 'sc-text-special sc-m-2x';
+  message.textContent = `Spinbox: Could not find playlist ${playlistName}`;
+  return message;
 }